Move all Previous page components into their own folder
