+++ src/components/previous/previousLayout.js
@@ -3,25 +3,25 @@ It will use the prevRevue.module.css module to format images and data.
 Can keep calling this to call all the info. This .js file will also store all the info of prev revues in it. */
 
 import React, {Component} from "react"
-import styles from "../components/previousLayout.module.css"
+import styles from "./previousLayout.module.css"
 
-import revue2019 from "../assets/images/previous/2019.jpg"
-import revue2018 from "../assets/images/previous/2018.jpg"
-import revue2017 from "../assets/images/previous/2017.jpg"
-import revue2016 from "../assets/images/previous/2016.jpg"
-import revue2015 from "../assets/images/previous/2015.jpg"
-import revue2014 from "../assets/images/previous/2014.jpg"
-import revue2013 from "../assets/images/previous/2013.jpg"
-import revue2012 from "../assets/images/previous/2012.jpg"
-import revue2011 from "../assets/images/previous/2011.jpg"
-import revue2010 from "../assets/images/previous/2010.jpg"
-import revue2009 from "../assets/images/previous/2009.gif"
-import revue2008 from "../assets/images/previous/2008.gif"
+import revue2019 from "../../assets/images/previous/2019.jpg"
+import revue2018 from "../../assets/images/previous/2018.jpg"
+import revue2017 from "../../assets/images/previous/2017.jpg"
+import revue2016 from "../../assets/images/previous/2016.jpg"
+import revue2015 from "../../assets/images/previous/2015.jpg"
+import revue2014 from "../../assets/images/previous/2014.jpg"
+import revue2013 from "../../assets/images/previous/2013.jpg"
+import revue2012 from "../../assets/images/previous/2012.jpg"
+import revue2011 from "../../assets/images/previous/2011.jpg"
+import revue2010 from "../../assets/images/previous/2010.jpg"
+import revue2009 from "../../assets/images/previous/2009.gif"
+import revue2008 from "../../assets/images/previous/2008.gif"
 
-import revue2019Cast from "../assets/images/previous/2019/cast_2019.jpg"
-import revue2018Cast from "../assets/images/previous/2018/cast_2018.jpg"
-import revue2017Cast from "../assets/images/previous/2017/cast_2017.jpg"
-import revue2016Cast from "../assets/images/previous/2016/cast_2016.jpg"
+import revue2019Cast from "../../assets/images/previous/2019/cast_2019.jpg"
+import revue2018Cast from "../../assets/images/previous/2018/cast_2018.jpg"
+import revue2017Cast from "../../assets/images/previous/2017/cast_2017.jpg"
+import revue2016Cast from "../../assets/images/previous/2016/cast_2016.jpg"
 
 class PreviousShow extends Component {
     constructor(props) {
